requirements.txt
streamlit>=1.45.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.24.0
requests>=2.32.0
beautifulsoup4>=4.12.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
schedule>=1.2.0
openpyxl>=3.1.0
xlrd>=2.0.0
scikit-learn>=1.5.0
